End and Bug Fix Commit
Added End screen and coupon for final screen after final hand. Fixed the sticky bug, was passing a card object to be created as a card onject rather than a string when the random number generated the same number as the first hand. Fixed by adding a while loop to check the new random against the old random which is saved before the reset. As a back up the card also ensures a string is the data type being passed on creation, if not it returns the parameter as is.